upd (*) reprojection improved 'as a service' + remove toats reprojecting step: useless now
